Added style to cookies
