+++ src/components/cookies/Cookies.css
@@ -1,0 +1,24 @@
+.container {
+  display: flex;
+  justify-content: center;
+}
+
+.cookie-container {
+  background-color: white;
+  width: 50vh;
+  height: 15vh;
+  padding: 10px;
+  border-radius: 5px;
+  text-align: center;
+  position: absolute;
+  top: 45vh;
+  box-shadow: 12px 12px 2px 1px rgba(36, 36, 37, 0.2);
+}
+
+.button {
+  border: none;
+  border-radius: 5px;
+  padding: 5px 10px;
+  margin: 0 10px;
+  background-color: rgba(36, 36, 37, 0.2);
+}
